numeroissime modifiche

+++ src/main/java/it/cleverad/engine/persistence/model/service/Budget.java
@@ -22,7 +22,7 @@ public class Budget {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Long budget;
+    private Double budget;
     @Column(name = "due_date")
     private LocalDate dueDate;
     @Column(name = "start_date")
