+++ src/main/java/it/cleverad/engine/persistence/model/service/Campaign.java
@@ -41,7 +41,7 @@ public class Campaign {
     private String trackingCode;
     @Column(name = "encoded_id")
     private String encodedId;
-    private Long budget;
+    private Double budget;
     @Column(nullable = false)
     private Boolean status = true;
     @Column(name = "creation_date")
